restaurant-backend/restaurant_dj/requirements.txt
amqp==5.1.0
appnope==0.1.2
asgiref==3.4.1
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.0
Deprecated==1.2.13
Django==4.0.3
django-cors-headers==3.10.1
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.2
django-nine==0.2.5
djangorestframework==3.13.1
elastic-transport==8.0.1
elasticsearch==7.13.4
elasticsearch-dsl==7.4.0
executing==0.8.2
ipython==7.30.1
jedi==0.18.1
kombu==5.2.3
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.28
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.1
Pygments==2.11.2
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
redis==4.1.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.1.3
traitlets==5.1.1
urllib3==1.26.7
uWSGI==2.0.20
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
